Remove ginkgo_mpi library, move to core
