+++ core/mpi/exception.cpp
