+++ core/mpi/get_info.cmake
