+++ core/test/CMakeLists.txt
@@ -2,6 +2,9 @@ include(${PROJECT_SOURCE_DIR}/cmake/create_test.cmake)
 
 add_subdirectory(accessor)
 add_subdirectory(base)
+if(GINKGO_BUILD_MPI)
+    add_subdirectory(mpi)
+endif()
 add_subdirectory(factorization)
 add_subdirectory(log)
 add_subdirectory(matrix)
