+++ core/test/mpi/CMakeLists.txt
